container_io: take read_loop function
